ai-service/requirements.txt
# ML/AI Libraries
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.2
spacy>=3.6.0
nltk>=3.8.1

# Google Gemini API (for native PDF processing)
google-generativeai>=0.3.0

# Core FastAPI
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0

# Image Processing
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0

# PDF Processing
PyPDF2>=3.0.0
pdf2image>=1.17.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Model Persistence
joblib>=1.3.0
pickle5>=0.0.12

# OpenAI (for hybrid approach)
openai>=1.0.0
